Added new functions, increased readability.
Two new functions were added, displaying the current unique id and
current unique views for the current logged in session. These are
illustrated in the file usage.php and the readme file. The code was
cleaned from all original *useless* comments, and indentation was added
to increase the readability of the library.